CampusHub-main/CampusHub-main/backend/requirements.txt
fastapi[standard]
uvicorn[standard]
motor
pymongo
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic[email]
pydantic-settings
google-generativeai
gunicorn
uvicorn[standard]
